Require Member data to populate SessionStore during creation
Previously the setMember call during create() was breaking application
of the method. This has been tweaked so that the setMember call is
first. Additionally, if there is no Member available, a blank
SessionState instance will always be created.

The StoreInterface has also been expanded to encompass more of the
functionality in SessionStore.