NEW add MFA protection to the reset password flow
When a logged in user vists e.g. Security/changepassword (default
route) they are prompted for their existing password and a new one.
However when they have forgotten their password and used
Security/forgotpassword to send an email with a reset link to the
holder of the account linked with that email address (if it exists)
then they do not have to enter the existing password first.

If MFA is installed AND enabled, then the person who visits that link
should be prompted to pass through an MFA check before the password
reset is allowed.