NEW event triggered notification system
Adds an interface for plugging and delivering notifications on events that have happened in the system.
These events are triggered by a user (which may or may not be null to indicate the system itself), and passed to a delivery manager for dispatch. An event can have multiple handlers, which are responsible for delivering notifications. By default there is an Email handler - but it's feasible to see the future addition of e.g. SMS, or perhaps a push notification to an app.

This is configured currently for:

- adding an MFA method (e.g. registration)
- removing an MFA method
- removing all MFA methods (reverting to single factor/password authentication)
- using a backup code for verification