Adding change password MFA interjection tests
Recently MFA was added to the "reset password" flow a user goes through
when activating the "forgot password" functionality on a site. In order
to assure this is the only affected flow (i.e. the normal change of a
password while logged in to the site is not affected), and only exists
when a user has MFA enabled, functional tests have been added to cover
these cases.