Finish rename of "LoginHandler" to "VerifyHandler"
As this better reflects the action taken by the handler. A number of tests and other areas were still referencing the old "LoginHandler" name as part of the MFA module's set of interfaces. Unchanged in this commit is the equally named "LoginHandler" to do with the SilverStripe v4 Security system - this class remains to reflect its relationship to the core class of the same name (it is not related to MFA Methods, rather the login form itself).