Add tests for new notification hooks
Recently added were notification events generated at certain points in the code.
These extensible points have already been prevented from sending notifications
while the system is under test - but they are not actively tested to be working.

This commit adds mocks and spies to ascertain that the notification system is
functional in the manner we expect in the appropriate places where they events
are generated to be sent.