FIX show correct error message and style to stand out
Issues with MFA verification were previously hard coded, meaning the error shown was not necessarily the error reported. This could lead to a user e.g. attempting to input the correct code to verify repeatedly, and be told that the code was incorrct - when in actuality it was that they were locked out of their account for a limited time due to too many incorrect attempts to verify.

Correcting this issue still left the error message not displayed in a manner that stood out from the rest of the interface, so a colour that is accessible to AAA contrast compliance has also been added in order to try and make the error message stand out from the rest of the interface, and thus hopefully be more noticable when it changes content.