FIX ensure RegisteredMFAFieldList always has a Member
and never "not a Member".

It was found that upon submission of a Member form in the CMS (e.g. to update a password)
there was a 500 server error being returned. The submission is of course attempting to
populate all the form fields with submitted data - the React form field for MFA management
is not a form field in the "accepts user input" sense, it is an interactive component
designed to perform actions directly - thus it has no value. No value means the PHP side
FormField instance is attempted to have `null` set as the value upon submission, which
disagrees strongly with the type hint of `Member`, and caused issues.

As this form field has two valid states - on where editing is possible and the member object
is always equal to `Security::getCurrentUser`, and a read only mode where the member may be any
valid member from the database - which means that simply relying on `Security::getCurrentUser`
would be unreliable, we must both require a member to always be set (via the constructor) and
not accept invalid `setValue` requests - but we cannot throw an exception here because this is
expected behaviour upon a form submission.