API alter user help link default translations
In order to be able to facilitate help text links that are both adaptive to their authenticator and translated,
a new API was added to translate this help link text on the server, and pass this through to the components via
the 'available method' schema. Once the method component takes over (is rendered within the MFA container) translations
again revert to javascript provided context.

Consistency was also created in how these help links were looked up from config.