increase accessiblity for unsupported state of method selection for registration
It was deemed previously that although contrast standards were being met, the text overlaying an image was too busy for the eye and was more difficult to read than it need be. A simplified design was made that also allowed the user to still access the "more info" link for a method as well.