FIX Email correct content via translation key correction
Upon registering a user is being sent an email to notify them that they have successfully added a multiple factor of authentication to their account on the site.
However due to a mix up with translation keys, the default text was being replaced with that for removing an authentication factor.
This commit fixes the translation key, and extends the tests to cover the body as well as the subject.