NEW Add a prop to disallow registration screens from showing in the modal
You can indicate which registration screens should not be accessible in the modal. Right now this is used to toggle away the modal if the user manages to navigate to the given screens. RegisteredMFAMethodListField has been updated to use this prop instead of it's own handling for this instance, and the Reset component now uses this prop to avoid showing invalid screens