FIX IE11 css grid issues with verify more options layout
IE11 follows an older version of the CSS Grid specification, which does now allow for implicit grid layout.
To this end the content and image section of the "more options" section of the MFA Verification flow were
laid out on top of each other, making it difficult/impossible to select another method to verify with.
By explicitly stating which column each element belongs to we solve this problem.