FIX IE11 inability to download backup codes
When registering for MFA the recovery method is the BackupCode method bundled with the module.
After these are registered the user is presented with them on screen, with options to print, download (save as), or copy them to the clipboard.
The download option was not functioning in IE 11 - interesting it was possible to create the block and generate the URL for it, however attempting to open the URL resulted in an error (only when manually testing in the console, normal output was non-existant upon clicking the button).

This _may_ be something to do with cross origin request bugs incorrectly identifying the blob URI as a different origin, an issue present in IE10 and apparently an issue in IE11 according to this comment, referencing an issue that is no longer accessible (since the bug tracker was retired by Microsoft): https://stackoverflow.com/questions/3665115/how-to-create-a-file-in-memory-for-user-to-download-but-not-through-server#comment43455460_20194533

This is all by the bye though, as the fix was simply to test for the IE specific blob save method and use that if present. This is present in Edge too, but was tested and did not seem to affect expected results (neither did this change alter any other non-MS browser behaviour).