FIX Check whether user is eligible for MFA in canSkipMFA method
Users with no CMS access are not eligible to use MFA by default. This
was respected in the shouldRedirectToMFA method, but not in canSkipMFA,
which resulted in these users being booted back to the login screen
whenever they attempted to log in.

This commit also tidies some of the docblocks in EnforcementManager, and
adjusts the hasAdminAccess method to always act as the provided Member.