FIX Drop already-verified check from SessionStore
While ostensibly a useful safety check, this validation step is
misplaced. The purpose of SessionStore is to hold context for the
verification process, not to verify its integrity, and there exist
(extremely hard-to-reproduce) edgecases where this will unhelpfully
break the MFA flow.

The MFA method verification process is sufficiently idempotent and
safe to prevent any negative outcomes from a method being verified
multiple times in a single MFA flow.